Checking function exists first

+++ header.php
@@ -194,25 +194,20 @@
     <!--<div class="container">-->
       <!--<div class="jumbotron">-->
         <!--<h1>Jumbotron!</h1>-->
-        <!--<p>-->
-          <!--This massive space is reserved for a featured post. We'll get this finished up before release.-->
-          <!--When finished, users can mark a post as 'featured' and it will be embiggened and put here automatically.-->
-        <!--</p>-->
       <!--</div>-->
     <!--</div>-->
 
     <?php
     // if we have an alert
-    if (stswr_alerts_get_current_alert()) {
-      if (stswr_alerts_get_current_alert('id') !== '0') {
+    if (function_exists('stswr_alerts_get_current_alert') && stswr_alerts_get_current_alert('id') !== '0') {
         echo '<div class="container alerts">';
         echo '<h1>'.stswr_alerts_get_current_alert('title').'</h1>';
         echo stswr_alerts_get_current_alert('body-html');
         echo '</div>';
-      } 
-    }
+    } 
+
     // if there is no alert, but a header image
-    elseif (get_header_image()) { ?>
+    else if (get_header_image()) { ?>
       <div class="container">
         <img src="<?php header_image(); ?>" height="<?php echo get_custom_header()->height; ?>" width="<?php echo get_custom_header()->width; ?>" alt="" />
       </div>
